Refactored event handling for improved readability and maintainability.

+++ src/extension/diagnostics/Diagnostic.ts
@@ -9,6 +9,7 @@ export class Diagnostic {
     if (document.languageId !== "jp1ajs") {
       return;
     }
+    console.log("checkForErrors");
     const result = parseAjs(document.getText());
     const diagnostics = result.errors.map((result) => {
       const startPos = new vscode.Position(
@@ -30,20 +31,30 @@ export class Diagnostic {
   }
 
   public static init(context: vscode.ExtensionContext) {
-    console.info("initialize Diagnostic.");
+    console.log("initialize Diagnostic.");
     const diagnostic = new Diagnostic();
     context.subscriptions.push(
-      vscode.workspace.onDidOpenTextDocument(diagnostic.checkForErrors),
-    );
-    context.subscriptions.push(
-      vscode.workspace.onDidChangeTextDocument((event) =>
-        diagnostic.checkForErrors(event.document),
-      ),
-    );
-    context.subscriptions.push(
-      vscode.workspace.onDidCloseTextDocument((doc) =>
-        diagnostic.diagnosticCollection.delete(doc.uri),
-      ),
+      vscode.workspace.onDidOpenTextDocument((doc) => {
+        if (doc.languageId !== "jp1ajs") {
+          return;
+        }
+        console.log("invoke Diagnostic.onDidOpenTextDocument.", doc);
+        diagnostic.checkForErrors(doc);
+      }),
+      vscode.workspace.onDidChangeTextDocument((event) => {
+        if (event.document.languageId !== "jp1ajs") {
+          return;
+        }
+        console.log("invoke Diagnostic.onDidChangeTextDocument.", event);
+        diagnostic.checkForErrors(event.document);
+      }),
+      vscode.workspace.onDidCloseTextDocument((doc) => {
+        if (doc.languageId !== "jp1ajs") {
+          return;
+        }
+        console.log("invoke Diagnostic.onDidCloseTextDocument.", doc);
+        diagnostic.diagnosticCollection.delete(doc.uri);
+      }),
     );
   }
 }
